backend/requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
click==8.2.1
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.6
daphne==4.2.1
Django==5.2.5
django-ratelimit==4.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
h11==0.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dotenv==1.1.1
redis==6.4.0
service-identity==24.2.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
whitenoise==6.9.0
zope.interface==7.2
